Corrected the tracking issue
Correction of commit f5fe586e979629ba757f245e5ecdf9e6bbdf6c05,
in which I substantialy changed the whole tracker workflow.
This is a patch that corrects the issue with wrong tracking,
which happened because I used wrong value of scale to
compute weighted max response in response map.